% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{df_component_plot}
\alias{df_component_plot}
\title{Extract selected variables}
\usage{
df_component_plot(object, select = NULL)
}
\arguments{
\item{object}{a dataframe or \link[ICS:ICS-S3]{ICS::ICS-S3} class object}

\item{select}{a vector of indexes of variables to plot. If \code{NULL} or
\code{FALSE}, all variables are selected. If \code{TRUE} only the first and last three are considered.}
}
\value{
A dataframe.
}
\description{
Extract the selected variables for a given dataframe or an invariant
coordinate system obtained via an ICS transformation.
}
