% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{df_select_plot}
\alias{df_select_plot}
\title{Extract kurtosis values}
\usage{
df_select_plot(object, select = NULL, scale = FALSE, screeplot = TRUE)
}
\description{
Extract the kurtosis values for an invariant
coordinate system obtained via an ICS transformation or for an object of
class 'ICS_crit'.
}
