% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_crit.R
\name{discriminatory_crit.default}
\alias{discriminatory_crit.default}
\title{Selection of ICS components based on discriminatory power}
\usage{
\method{discriminatory_crit}{default}(
  object,
  clusters,
  method = "eta2",
  nb_select = NULL,
  select_only = FALSE,
  gen_kurtosis = NULL
)
}
\arguments{
\item{object}{}

\item{clusters}{}

\item{method}{}

\item{nb_select}{}

\item{select_only}{}
}
\description{
This function selects the components to keep based on a discriminatory power
(by default "eta2")
}
