% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_crit.R
\name{fixOrder}
\alias{fixOrder}
\title{Fix Order based on rolling variances}
\usage{
fixOrder(x, nb_spherical)
}
\arguments{
\item{x}{a vector of generalized kurtosis values}

\item{nb_spherical}{the number of non spherical components to consider.}
}
\value{
A list with:
\itemize{
\item \code{RollVarX}: the rolling variances.
\item \code{Order}: indexes of the ordered invariant components.
}
}
\description{
Simplified version of the non exported function \code{ICtest:::fixOrder()}.
}
