% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_comp_crit.R
\name{ICS_comp_crit}
\alias{ICS_comp_crit}
\title{Title}
\usage{
ICS_comp_crit(
  object,
  crit = "normal",
  crit_args = list(),
  threshold = NULL,
  groups_vec = NULL,
  ...
)
}
\arguments{
\item{object}{Object of class 'ICS'}

\item{crit}{One of teh following criteria: 'normal', 'med', 'var', 
'discriminatory'.}

\item{threshold}{The maximal number of components to select.}

\item{...}{}

\item{k}{The number of known clusters.}
}
\description{
Title
}
