% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatters.R
\name{ICS_rmcd}
\alias{ICS_rmcd}
\title{RMCD location and Scatter Estimates for ICS}
\usage{
ICS_rmcd(x, location = FALSE, nsamp = "deterministic", alpha = 0.5, ...)
}
\arguments{
\item{x}{a numeric matrix or data frame.}

\item{location}{a logical indicating whether to include the MCD-estimate of
location (defaults to \code{FALSE}).}

\item{nsamp}{number of subsets used for initial estimates or "best",
"exact" or "deterministic" (default).}

\item{alpha}{numeric parameter controlling the size of the subsets over
which the determinant is minimized as in \code{\link[rrcov:CovMcd]{rrcov::CovMcd()}}.}

\item{...}{potential further arguments passed to \code{\link[rrcov:CovMcd]{rrcov::CovMcd()}}.}
}
\value{
An object of class \code{"ICS_scatter"} with the following
components:
\item{location}{if requested, a numeric vector giving the location
estimate.}
\item{scatter}{a numeric matrix giving the estimate of the scatter matrix.}
\item{label}{a character string providing a label for the scatter matrix.}
}
\description{
It is a wrapper for the reweighted MCD estimator of location and scatter
as computed by \code{\link[rrcov:CovMcd]{rrcov::CovMcd()}}.
}
\details{
The raw estimates in combination of \code{nsamp}="deterministic" are available
in RForge \code{install.packages("robustbase",  repos = c("http://R-Forge.R-project.org", "http://cran.at.r-project.org"), dep = TRUE)}
}
\seealso{
\code{\link[rrcov:CovMcd]{rrcov::CovMcd()}}, \code{\link[=ICS_mcd]{ICS_mcd()}}.
}
