% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust-package.R
\docType{package}
\name{ICSClust-package}
\alias{ICSClust-package}
\title{\packageTitle{ICSClust}}
\description{
\packageDescription{ICSClust}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{ICSClust}
\packageIndices{ICSClust}
}
\references{
Alfons, A., Archimbaud, A., Nordhausen, K., & Ruiz-Gazen, A. (2024).
Tandem clustering with invariant coordinate selection.
Econometrics and Statistics.
\doi{10.1016/j.ecosta.2024.03.002}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AuroreAA/ICSClust}
  \item Report bugs at \url{https://github.com/AuroreAA/ICSClust/issues}
}

}
\author{
\packageAuthor{ICSClust}

Maintainer: \packageMaintainer{ICSClust}
}
\keyword{package}
