% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandemClust.R
\name{kmeansClust}
\alias{kmeansClust}
\title{Kmeans clustering}
\usage{
kmeansClust(df, k, iter.max = 100, nstart = 20, ...)
}
\arguments{
\item{iter.max}{}

\item{nstart}{}

\item{...}{}

\item{x}{}

\item{centers}{}
}
\description{
The output is of class tandemClust
}
