% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_crit.R
\name{normal_crit.default}
\alias{normal_crit.default}
\title{Normal criterion}
\usage{
\method{normal_crit}{default}(
  object,
  level = 0.05,
  test = c("agostino.test", "jarque.test", "anscombe.test", "bonett.test",
    "shapiro.test"),
  max_select = NULL,
  select_only = FALSE
)
}
\arguments{
\item{object}{}

\item{level}{}

\item{test}{}

\item{max_select}{}

\item{select_only}{}
}
\description{
Normal criterion
}
