% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust.R
\name{plot.ICSClust}
\alias{plot.ICSClust}
\title{Scatterplot Matrix with densities on the diagonal}
\usage{
\method{plot}{ICSClust}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"ICSClust"}.}

\item{\dots}{additional arguments to be passed down to \code{\link[=component_plot]{component_plot()}}}
}
\value{
An object of class \code{\link[GGally:ggmatrix]{"ggmatrix"}} (see
\code{\link[GGally:ggpairs]{GGally::ggpairs()}}).
}
\description{
Wrapper for \code{\link[=component_plot]{component_plot()}}.
}
\author{
Aurore Archimbaud
}
