% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust.R
\name{print.ICSClust_summary}
\alias{print.ICSClust_summary}
\title{Print of an \code{ICSClust_summary} object}
\usage{
\method{print}{ICSClust_summary}(x, info = FALSE, digits = 4L, ...)
}
\arguments{
\item{x}{object of class \code{"ICSClust_summary"}.}

\item{info}{logical, either TRUE or FALSE. If TRUE, prints additional
information on arguments used for computing scatter matrices
(only named arguments that contain numeric, character, or logical scalars)
and information on the parameters of the algorithm.
Default is FALSE.}

\item{digits}{number of digits for the numeric output.}

\item{...}{additional arguments are ignored.}
}
\value{
The supplied object of class \code{"ICSClust_summary"} is returned invisibly.
}
\description{
Prints an \code{ICSClust_summary} object in an informative way.
}
\author{
Aurore Archimbaud
}
