% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatters.R
\name{scov}
\alias{scov}
\title{One-step M-estimator of scatter}
\usage{
scov(x, beta = 0.2)
}
\arguments{
\item{x}{a numeric matrix or data frame.}

\item{beta}{a positive numeric value specifying the tuning parameter of the
one-step M-estimator (defaults to 0.2), see \sQuote{Details}.}
}
\value{
A numeric matrix giving the one-step M-estimate of scatter.
}
\description{
Compute a one-step M-estimator of scatter with weights based on Mahalanobis
distances.
}
\details{
For a sample \eqn{\boldsymbol{X}_{n} = (\mathbf{x}_{1}, \dots, 
\mathbf{x}_n)^{\top}}, a positive and decreasing weight function \eqn{w},
and a tuning parameter \eqn{\beta > 0}, the one-step M-estimator
of scatter is defined as
\deqn{\mathrm{SCOV}_\beta(\boldsymbol{X}_{n}) =
\frac{\sum_{i=1}^{n} 
w(\beta \, r^{2}(\mathbf{x}_{i})) 
(\mathbf{x}_{i} - \mathbf{\bar{x}}_{n}) 
(\mathbf{x}_{i} -  \mathbf{\bar{x}}_{n})^{\top}}{\sum_{i=1}^{n} 
w(\beta \, r^{2}(\mathbf{x}_{i}))},}
where
\deqn{r^{2}(\mathbf{x}_{i}) = 
(\mathbf{x}_{i} -  \mathbf{\bar{x}}_{n})^{\top}
\mathrm{COV}(\boldsymbol{X}_n)^{-1} 
(\mathbf{x}_{i} -  \mathbf{\bar{x}}_{n})}
denotes the squared Mahalanobis distance of observation \eqn{\mathbf{x}_{i}}
from the sample mean \eqn{\mathbf{\bar{x}}_{n}} based on the sample
covariance matrix \eqn{\mathrm{COV}(\boldsymbol{X}_n)}. Here, the weight
function \eqn{w(x) = \exp(-x/2)} is used.
}
\references{
Caussinus, H. and Ruiz-Gazen, A. (1993) Projection Pursuit and Generalized
Principal Component Analysis. In Morgenthaler, S., Ronchetti, E., Stahel,
W.A. (eds.) \emph{New Directions in Statistical Data Analysis and Robustness},
35-46. Monte Verita, Proceedings of the Centro Stefano Franciscini Ascona
Series. Springer-Verlag.

Caussinus, H. and Ruiz-Gazen, A. (1995) Metrics for Finding Typical
Structures by Means of Principal Component Analysis. In \emph{Data Science and
its applications}, 177-192. Academic Press.
}
\author{
Andreas Alfons and Aurore Archimbaud
}
