% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scree_plot.ICS}
\alias{scree_plot.ICS}
\title{Plot the Generalized Kurtosis Values of the ICS Transformation}
\usage{
\method{scree_plot}{ICS}(
  object,
  select = NULL,
  scale = FALSE,
  type = c("dots", "lines"),
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{"ICS"} containing
results from an ICS transformation.}

\item{select}{an integer, character, or logical vector specifying for which
components to extract the generalized kurtosis values, or \code{NULL} to
extract the generalized kurtosis values of all components.}

\item{scale}{a logical indicating whether to scale the generalized kurtosis
values to have product 1 (defaults to \code{FALSE}).}
}
\description{
Extract the generalized kurtosis values of the components obtained via an
ICS transformation and draw a screeplot.
}
