% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{select_plot.data.frame}
\alias{select_plot.data.frame}
\title{Plot the Generalized Kurtosis Values of the ICS Transformation}
\usage{
\method{select_plot}{data.frame}(df, type = c("dots", "lines"), select = NULL, ...)
}
\arguments{
\item{df}{}

\item{type}{}

\item{...}{}
}
\description{
Extract the generalized kurtosis values of the components obtained via an
ICS transformation and draw a screeplot.
}
