% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{select_plot.default}
\alias{select_plot.default}
\title{Title}
\usage{
\method{select_plot}{default}(
  df,
  select = NULL,
  scale = FALSE,
  int = 0.2,
  color = "grey",
  alpha = 0.3,
  screeplot = TRUE,
  ...
)
}
\arguments{
\item{df}{}

\item{select}{}

\item{scale}{}

\item{int}{}

\item{color}{}

\item{alpha}{}

\item{screeplot}{}

\item{...}{}
}
\description{
Title
}
