% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{select_plot.ICS_crit}
\alias{select_plot.ICS_crit}
\title{Title}
\usage{
\method{select_plot}{ICS_crit}(object, select = NULL, scale = FALSE, screeplot = TRUE, ...)
}
\arguments{
\item{object}{}

\item{select}{}

\item{scale}{}

\item{...}{}

\item{type}{}
}
\description{
Title
}
