% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust.R
\name{summary.ICSClust}
\alias{summary.ICSClust}
\title{Summary of an \code{ICSClust} object}
\usage{
\method{summary}{ICSClust}(object, ...)
}
\arguments{
\item{object}{object of class \code{"ICSClust"}.}

\item{...}{additional arguments passed to \code{\link[=summary]{summary()}}}
}
\value{
An object of class \code{"ICSClust_summary"} with the following components:
\itemize{
\item \code{ICS_out}: \code{ICS_out} object
\item \code{nb_comp}: number of selected components
\item \code{select}: vector of names of selected components
\item \code{nb_clusters}: number of clusters
\item \code{table_clusters}: frequency table of clusters
}
}
\description{
Summarizes an \code{ICSClust} object in an informative way.
}
\author{
Aurore Archimbaud
}
