% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_crit.R
\name{var_crit.default}
\alias{var_crit.default}
\title{Rollvar to select the ICS component}
\usage{
\method{var_crit}{default}(object, index_max = NULL, index_only = FALSE)
}
\arguments{
\item{object}{The vector of lambdas from ICS object}

\item{index_max}{The number of non-gaussian components under the null.}

\item{index_only}{}
}
\description{
This function selects the components to keep based on the rolling variance criteria used in `ICSboot`.
}
