% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_crit.R
\name{discriminatory_crit.default}
\alias{discriminatory_crit.default}
\title{Selection of ICS components based on discriminatory power}
\usage{
\method{discriminatory_crit}{default}(
  object,
  clusters,
  method = "eta2",
  index_max = NULL,
  index_only = FALSE
)
}
\arguments{
\item{object}{}

\item{clusters}{}

\item{method}{}

\item{index_max}{}

\item{index_only}{}
}
\description{
This function selects the components to keep based on a discriminatory power (by default "eta2")
}
