% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_comp_crit.R
\name{discriminatory_crit}
\alias{discriminatory_crit}
\title{Selection of ICS components based on discriminatory power}
\usage{
discriminatory_crit(
  df_scores,
  groups_vec,
  crit_args = list(method = "eta2"),
  comp_max = NULL
)
}
\arguments{
\item{df_scores}{}

\item{groups_vec}{}

\item{crit_args}{}

\item{comp_max}{}
}
\description{
This function selects the components to keep based on a discriminatory power (by default "eta2")
}
