% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatters.R
\name{ICS_mlc}
\alias{ICS_mlc}
\title{## reference implementation using package 'amap'
#' @importFrom amap W
#' @importFrom stats var
ucov_amap <- function(x, beta = 0.2) {
  # initializations
  x <- as.matrix(x)
  cn <- colnames(x)
  # compute inverse of maximum likelihood estimate of covariance matrix
  # n <- nrow(x)
  # S_inv <- solve(var(x) * ((n-1) / n))
  S_inv <- solve(var(x))
  # compute inverse of SCOV (error in 'amap' package: should be 1/h^2)
  h <- 1/sqrt(beta)
  scov_inv <- solve(amap::varrob(x, h = h, D = S_inv, kernel = "gaussien"))
  V <- solve(scov_inv + (h-1)/(h^2) * S_inv)
  # set row and column names and return scatter matrix
  dimnames(V) <- list(cn, cn)
  V
}}
\usage{
ICS_mlc(x, location = FALSE, ...)
}
\description{
## reference implementation using package 'amap'
#' @importFrom amap W
#' @importFrom stats var
ucov_amap <- function(x, beta = 0.2) {
  # initializations
  x <- as.matrix(x)
  cn <- colnames(x)
  # compute inverse of maximum likelihood estimate of covariance matrix
  # n <- nrow(x)
  # S_inv <- solve(var(x) * ((n-1) / n))
  S_inv <- solve(var(x))
  # compute inverse of SCOV (error in 'amap' package: should be 1/h^2)
  h <- 1/sqrt(beta)
  scov_inv <- solve(amap::varrob(x, h = h, D = S_inv, kernel = "gaussien"))
  V <- solve(scov_inv + (h-1)/(h^2) * S_inv)
  # set row and column names and return scatter matrix
  dimnames(V) <- list(cn, cn)
  V
}
}
