% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust-package.R
\docType{package}
\name{ICSClust-package}
\alias{ICSClust-package}
\title{\packageTitle{ICSClust}}
\description{
\packageDescription{ICSClust}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{ICSClust}
\packageIndices{ICSClust}
}
\references{
Alfons, A., Archimbaud, A., Nordhausen, K., & Ruiz-Gazen, A. (2022).
Tandem clustering with invariant coordinate selection.
\emph{arXiv preprint arXiv:2212.06108}.
}
\author{
\packageAuthor{ICSClust}

Maintainer: \packageMaintainer{ICSClust}
}
\keyword{package}
