% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust.R
\name{ICSClust}
\alias{ICSClust}
\title{Tandem clustering with ICS}
\usage{
ICSClust(
  X,
  ICS_args = list(),
  nb_select = NULL,
  criterion = c("med_crit", "normal_crit", "var_crit", "discriminatory_crit"),
  ICS_crit_args = list(),
  clusters = NULL,
  nb_clusters = NULL,
  method = c("kmeans_clust", "tkmeans_clust", "pam_clust"),
  clustering_args = list()
)
}
\arguments{
\item{X}{}

\item{ICS_args}{}

\item{nb_select}{}

\item{criterion}{}

\item{ICS_crit_args}{}

\item{clusters}{}

\item{nb_clusters}{}

\item{method}{}

\item{clustering_args}{}
}
\description{
Tandem clustering with ICS
}
