% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust.R
\name{ICSClust}
\alias{ICSClust}
\title{Title}
\usage{
ICSClust(
  object,
  crit = "normal",
  crit_args = list(),
  groups_vec = NULL,
  comp_max = NULL,
  scaling = FALSE,
  clustering_fun = "kmeansClust",
  clustering_args = list(),
  nb_clusters = NULL,
  ...
)
}
\arguments{
\item{object}{Object of class 'ICS'}

\item{...}{}
}
\description{
Title
}
