% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust.R
\name{ICSClust}
\alias{ICSClust}
\title{Tandem clustering with ICS}
\usage{
ICSClust(
  X,
  nb_select = NULL,
  nb_clusters = NULL,
  ICS_args = list(),
  criterion = c("med_crit", "normal_crit", "var_crit", "discriminatory_crit"),
  ICS_crit_args = list(),
  method = c("kmeans_clust", "tkmeans_clust", "pam_clust"),
  clustering_args = list(),
  clusters = NULL
)
}
\arguments{
\item{X}{a numeric matrix or data frame containing the data.}

\item{nb_select}{the number of components to select. \code{NULL} by default.
It is required in case \code{criterion} is either "med_crit", "var_crit" or
"discriminatory_crit".}

\item{nb_clusters}{the number of clusters searched for.}

\item{ICS_args}{list of \code{\link[ICS]{ICS-S3}} arguments. Otherwise, default
values of \code{\link[ICS]{ICS-S3}} are used.}

\item{criterion}{criterion to automatically decide which invariant components
to keep. Possible values are "med_crit", "normal_crit", "var_crit" and
"discriminatory_crit". The default value is "med_crit".
See below for more details.}

\item{ICS_crit_args}{list of \code{\link[=med_crit]{med_crit()}}, \code{\link[=normal_crit]{normal_crit()}}, \code{\link[=var_crit]{var_crit()}} or
\code{\link[=discriminatory_crit]{discriminatory_crit()}}, arguments for choosing the components to keep.}

\item{method}{clustering method to perform. Possible implemented wrapper
functions are named "kmeans_clust", "tkmeans_clust" or "pam_clust".
The default value is "kmeans_clust".}

\item{clustering_args}{list of \code{\link[=kmeans_clust]{kmeans_clust()}},
\code{\link[=tkmeans_clust]{tkmeans_clust()}} or \code{\link[=pam_clust]{pam_clust()}} arguments for performing cluster analysis.}

\item{clusters}{a vector indicating the true clusters of the data. By default,
it is \code{NULL} but it is required to choose the components based on the
discriminatory criterion \code{\link{discriminatory_crit}}.}
}
\value{
An object with the following components:
\itemize{
\item ICS_out: An object of class \code{"ICS"}.
See \code{\link[=ICS-S3]{ICS}}
\item select: a vector of the names of the selected invariant
coordinates.
\item clusters: a vector of the new partition of the data, i.e a vector
of integers (from \code{1:k}) indicating the cluster to which each
observation is allocated. 0 indicates outlying observations.
}
}
\description{
Sequential clustering approach: (i) dimension reduction through the Invariant
Coordinate Selection method using the \code{\link[=ICS-S3]{ICS}} function and (ii)
clustering of the transformed data.
}
\details{
Tandem clustering with ICS is a sequential method:
\itemize{
\item \code{\link[=ICS-S3]{ICS}} is performed
\item only a subset of the first and/or the last components are
selected based on a criterion.
\item the clustering method is performed only onto the subspace
of the selected components.
}
}
\examples{
\dontrun{
X <- iris[,1:4]

# indicating the number of components to retain for the dimension reduction
step as well as the number of clusters searched for.
out <- ICSClust(X, nb_select = 2, nb_clusters = 3)

# changing the scatter pair to consider in ICS
out <- ICSClust(X, nb_select = 1, nb_clusters = 3,
 ICS_args = list(S1 = ICS_mcd, S2 = ICS_cov,S1_args = list(alpha = 0.5)))
 
# changing the criterion for choosing the invariant coordinates
out <- ICSClust(X, nb_clusters = 3, criterion = "normal_crit",
ICS_crit_args = list(level = 0.1, test = "anscombe.test", max_select = NULL))

# changing the clustering method
out <- ICSClust(X, nb_clusters = 3, method  = "tkmeans_clust",
 clustering_args = list(alpha = 0.1))
 }
}
\references{
Alfons, A., Archimbaud, A., Nordhausen, K., & Ruiz-Gazen, A. (2022).
Tandem clustering with invariant coordinate selection.
\emph{arXiv preprint arXiv:2212.06108}.
}
\seealso{
\code{\link[=med_crit]{med_crit()}}, \code{\link[=normal_crit]{normal_crit()}},
\code{\link[=var_crit]{var_crit()}}, \link[=ICS-S3]{ICS},
\code{\link[=discriminatory_crit]{discriminatory_crit()}}, \code{\link[=kmeans_clust]{kmeans_clust()}}, \code{\link[=tkmeans_clust]{tkmeans_clust()}} and
\code{\link[=pam_clust]{pam_clust()}}
}
\author{
Andreas Alfons and Aurore Archimbaud
}
