% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kmeans_clust}
\alias{kmeans_clust}
\title{Kmeans clustering}
\usage{
kmeans_clust(df, k, clusters_only = FALSE, iter.max = 100, nstart = 20, ...)
}
\arguments{
\item{df}{}

\item{k}{}

\item{clusters_only}{}

\item{iter.max}{}

\item{nstart}{}

\item{...}{}
}
\description{
Kmeans clustering
}
