% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{mixture_sim}
\alias{mixture_sim}
\title{Simulation of mixture of gaussian distributions}
\usage{
mixture_sim(pct_clusters = c(0.5, 0.5), n = 500, p = 10, delta = 10)
}
\arguments{
\item{pct_clusters}{}

\item{n}{}

\item{p}{}

\item{delta}{}

\item{sigma}{}

\item{mean_grp}{}

\item{outliers}{}

\item{eps}{}
}
\description{
Simulation of mixture of gaussian distributions
}
