% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_crit.R
\name{normal_crit.default}
\alias{normal_crit.default}
\title{Title}
\usage{
\method{normal_crit}{default}(
  object,
  level = 0.05,
  test = c("jarque.test", "anscombe.test", "bonett.test", "agostino.test",
    "shapiro.test"),
  index_max = NULL,
  index_only = FALSE
)
}
\arguments{
\item{object}{}

\item{level}{}

\item{test}{}

\item{index_max}{}

\item{index_only}{}
}
\description{
Title
}
