% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_comp_crit.R
\name{normal_crit}
\alias{normal_crit}
\title{Title}
\usage{
normal_crit(
  df_scores,
  crit_args = list(level = 0.05, test = "agostino.test"),
  threshold = NULL
)
}
\arguments{
\item{crit_args}{}

\item{threshold}{}
}
\description{
Title
}
