% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pam_clust}
\alias{pam_clust}
\title{Title}
\usage{
pam_clust(df, k, clusters_only = FALSE, diss = inherits(df, "dist"), ...)
}
\arguments{
\item{df}{}

\item{k}{}

\item{clusters_only}{}

\item{diss}{}

\item{..}{}
}
\description{
Title
}
