% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{runif_outside_range}
\alias{runif_outside_range}
\title{Outside range}
\usage{
runif_outside_range(n, min = 0, max = 1, mult = 2)
}
\arguments{
\item{n}{number of observations to generate.}

\item{min}{numeric vector giving the minimum of each variable of the data set.}

\item{max}{numeric vector giving the maximum of each variable of the data set.}

\item{mult}{multiplication factor to expand the hyperrectangle around the data
(which is given by 'min' and 'max'), e.g., the default value 2 gives
a hyperrectangle for which each side is twice as long.  The data are
then drawn from a uniform distribution on the expanded hyperrectangle
from which the smaller hyperrectangle around the data is cut out.}
}
\value{
A matrix of generated points.
}
\description{
Draw from a multivariate uniform distribution outside range of a data set
}
\author{
Andreas Alfons
}
