% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSClust.R
\name{summary.ICSClust}
\alias{summary.ICSClust}
\title{Summarize an \code{ICSClust} object}
\usage{
\method{summary}{ICSClust}(object, info = FALSE, digits = 4L, ...)
}
\arguments{
\item{object}{object of class \code{ICSClust}.}

\item{info}{Logical, either TRUE or FALSE. If TRUE, print additional
information on arguments used for computing scatter matrices
(only named arguments that contain numeric, character, or logical scalars)
and information on the parameters of the algorithm.
Default is FALSE.}

\item{digits}{number of digits for the numeric output.}

\item{...}{additional arguments are ignored.}
}
\description{
Summarizes and prints an \code{ICSClust} object in an informative way.
}
\author{
Aurore Archimbaud
}
