% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_crit.R
\name{var_crit.default}
\alias{var_crit.default}
\title{Rollvar to select the ICS component}
\usage{
\method{var_crit}{default}(object, nb_select = NULL, select_only = FALSE)
}
\arguments{
\item{object}{The vector of gen_kurtosis from ICS object}

\item{nb_select}{The number of non-gaussian components under the null.}

\item{select_only}{}
}
\description{
This function selects the components to keep based on the rolling variance
criteria used in \code{ICSboot}.
}
