% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_crit.R
\name{var_crit}
\alias{var_crit}
\alias{var_crit.ICS}
\alias{var_crit.default}
\title{Selection of Invariant components using the var criterion}
\usage{
var_crit(object, ...)

\method{var_crit}{ICS}(object, ...)

\method{var_crit}{default}(object, nb_select = NULL, select_only = FALSE)
}
\arguments{
\item{object}{object of class \code{ICS}.}

\item{...}{not used.}

\item{nb_select}{the exact number of components to select.}

\item{select_only}{boolean. If \code{TRUE} only the vector names of the selected
invariant components is returned. If \code{FALSE} additional details are returned.}
}
\value{
If \code{select_only} is \code{TRUE} a vector of the names of the invariant
components or variables to select. If \code{FALSE} an object of class \code{ICS_crit}
is returned with the following objects:
\itemize{
\item \code{crit}: the name of the criterion "var".
\item \code{nb_select}: the number of components to select.
\item \code{gen_kurtosis}: the vector of generalized kurtosis values.
\item \code{select}: the names of the invariant components or variables to select.
\item \code{RollVarX}: the rolling variances.
\item \code{Order}: indexes of the ordered invariant components.
}
}
\description{
Identifies non spherical invariant coordinates based on the rolling variance
criterion used in \code{\link[ICtest:ICSboot]{ICtest::ICSboot()}}.
}
\examples{
\dontrun{
X <- iris[,-5]
out <- ICS(X)
var_crit(out, nb_select = 2, select_only = FALSE)
}
}
\references{
Alfons, A., Archimbaud, A., Nordhausen, K., & Ruiz-Gazen, A. (2022).
Tandem clustering with invariant coordinate selection.
\emph{arXiv preprint arXiv:2212.06108}.
}
\seealso{
\code{\link[=normal_crit]{normal_crit()}}, \code{\link[=med_crit]{med_crit()}}, \code{\link[=discriminatory_crit]{discriminatory_crit()}},
\code{\link[ICtest:ICSboot]{ICtest::ICSboot()}}.
}
\author{
Andreas Alfons, Aurore Archimbaud and Klaus Nordhausen
}
