% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_comp_crit.R
\name{var_crit}
\alias{var_crit}
\title{Rollvar to select the ICS component}
\usage{
var_crit(lambda_vec, comp_max = NULL)
}
\arguments{
\item{lambda_vec}{The vector of lambdas from ICS object}

\item{comp_max}{The number of non-gaussian components under the null.}
}
\description{
This function selects the components to keep based on the rolling variance criteria used in `ICSboot`.
}
